nvapi: Use __func__ macro for function names.
Improves logging a bit by removing hard-coded strings for function
names.

Uses __func__ due to its being portable and accessible in all the
compilers after 90s, while __FUNCTION__ isn't portable and is a gnu
extension, and we know we build the project with more or less modern
compilers. The __PRETTY_FUNCTION__ isn't used as we only need to have a
function name printed out.