nvapi-disp,sys: Build unique display IDs
We used to treat the display ID like an index, but this can be confusing
since the first display does not has display ID 0. Building a unique ID
is more like this works on Windows. The IDs are of course different on
Windows, but this shouldn't be an issue since those ID's are used like
an opaque token.  The former implementation didn't matches IDs on
Windows either.

Also some renames and make index type uint32_t to match the nvapi type.