d3d12: Prepare for ID3D12GraphicsCommandListExt1
Introduce a new struct NvapiD3d12Device::CommandListExtWithVersion
allowing for pointer compatibility betweeen revisions of the
ID3D12GraphicsCommandListExt interface (like
ID3D12GraphicsCommandListExt1 which will be introduced in a follow-up
change).

Alter the NvapiD3d12Device::GetCommandListExt() function to return this
structure. This will allow a single call to the function for returning
the highest available interface version for the current command list.
Callers of this function must check the interface version in order to
ensure they are only calling functions which are supported by the
interface.