tests: Make mocks compatible with MSVC COM return ABI
The canonical ABI for COM method calls that return a structure by value is
to "pass a hidden first parameter that is the address of an uninitialized
structure in which to place the result" [1]. As GCC does not conform to
this convention, MinGW goes through length to workaround the ABI issue by
adding a wrapper that explicitly takes the hidden return value parameter.

This hack is unnecessary on MSVC. Unfortunately, it also means that we need
to declare mocks with different argument count. Use ifdef against
WIDL_EXPLICIT_AGGREGATE_RETURNS (the macro defined by MinGW to enable the
wrappers) to switch to the appropriate version. For MSVC, there are no
overloads coming from the ABI hack, so we can use the simpler
IMPLEMENT_MOCK macros here.

The luid return was also technically wrong (it should write to the pointer
argument, not return an unrelated address), but the bug was masked by the
MinGW wrapper. Fix it, just in case.

[1]: https://devblogs.microsoft.com/oldnewthing/20220113-00/?p=106152