layer: Relax Meson requirement from 1.1 to 1.0
Meson 1.0 is an interesting target because it's the version provided by
the Debian 12 stable release, and by Valve's Steam Runtime (version 2 and
up), therefore indirectly by Proton. The only obvious use of a Meson 1.1
feature in `layer/` is the `meson.options` name, which is easily avoided.

In particular, the layer can be successfully compiled in the
Steam Runtime 3 'sniper' SDK container[1], using its backport of gcc 14:

    $ sudo vi /etc/apt/sources.list    # add the sniper_beta suite
    $ sudo apt install gcc-14-monolithic
    $ CC=gcc-14 CXX=g++-14 meson setup _build-sniper
    $ meson compile -C _build-sniper

[1] https://gitlab.steamos.cloud/steamrt/sniper/sdk

Signed-off-by: Simon McVittie <smcv@collabora.com>